Refactor diagnostics.  The xfel processing script will now output diagnostics to the out/debug folder in this format: hostname,event timestamp,now,status,detail
Where hostname is the machine being run on, event timestamp is the timestamp of the event being processed, now is the current time, status is either "    ", done, fail, or stop depending on processing status and detail is extra information as available.  The program will log when it starts a new task and when it is about to return for each event.  If the event successfully processes, the status will be done.  If there is an unexpected failure, the status will be fail. If the processing fails in an expected way (not enough spots, couldn't index), the status will be stop. If the whole program crashes unexpectedly, the status will likely be "    ", meaning it crashed in the middle of processing after emitting a 'task started' log entry.